Change level for "Reporting span" message to DEBUG
This library is really chatty, creating a few line of logs for every
event, even in prod. This changes the level to DEBUG, and does not call
on `_pretty_span` and `_pretty_logs` if the output is not required (not
having the `if` there caused them to be called even if logging level was
greater than WARN).