Use TBinaryProtocolAccelerated when serializing reports.
We're currently seeing our services spend a large amount of time inside of the lightstep sdk
serializing reports when they instrument themselves with tracing.  This is in large part due to
the fact that lightstep is using the pure python implemenation of Thrift.

Fixes https://github.com/lightstep/lightstep-tracer-python/issues/89